# Replace Regular Expressions with custom hash parser