--- a/src/BCrypt.Net/HashParser.cs
+++ b/src/BCrypt.Net/HashParser.cs
@@ -6,36 +6,48 @@ namespace BCrypt.Net
     {
         public static HashInformation GetHashInformation(string hash)
         {
-            if (hash == null)
+            if (!IsValidHash(hash))
+            {
+                ThrowInvalidHashFormat();
+            }
+
+            return new HashInformation(hash.Substring(0, 6), hash.Substring(1, 2), hash.Substring(4, 2), hash.Substring(7));
+        }
+
+        private static bool IsValidHash(string hash)
+        {
+            if (hash is null)
             {
                 throw new ArgumentNullException(nameof(hash));
             }
 
+            // Validate settings
             if (hash.Length != 60
                 || !hash.StartsWith("$2")
                 || !IsValidBCryptVersionChar(hash[2])
                 || hash[3] != '$'
                 || hash[6] != '$')
             {
-                ThrowInvalidHashFormat();
+                return false;
             }
 
-            string workFactorAttempt = hash.Substring(4, 2);
-            if (!int.TryParse(workFactorAttempt, out int workFactor)
-                || workFactor < 0)
+            // Validate workfactor
+            if (!IsAsciiNumeric(hash[4])
+                || !IsAsciiNumeric(hash[5]))
             {
-                ThrowInvalidHashFormat();
+                return false;
             }
 
+            // Validate hash
             for (int i = 7; i < hash.Length; ++i)
             {
                 if (!IsValidBCryptBase64Char(hash[i]))
                 {
-                    ThrowInvalidHashFormat();
+                    return false;
                 }
             }
 
-            return new HashInformation(hash.Substring(0, 6), hash.Substring(1, 2), workFactorAttempt, hash.Substring(7));
+            return true;
         }
 
         private static bool IsValidBCryptVersionChar(char value)
@@ -54,6 +66,11 @@ namespace BCrypt.Net
                 || (value >= '0' && value <= '9')
                 || (value >= 'A' && value <= 'Z')
                 || (value >= 'a' && value <= 'z');
+        }
+
+        private static bool IsAsciiNumeric(char value)
+        {
+            return value >= '0' && value <= '9';
         }
 
         private static void ThrowInvalidHashFormat()