Removed the dependency on regular expressions and added a quite simple hash parser.

Improves performance for `PasswordNeedsRehash` roughly 8x, while reducing memory usage by 6x.

Before:

|              Method |                 text |                 hash |                value |             Mean |         Error |        StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------------- |--------------------- |--------------------- |-----------------:|--------------:|--------------:|-------:|------:|------:|----------:|
| PasswordNeedsRehash |                    ? |                    ? |                    ? |       1,626.2 ns |      14.83 ns |      13.15 ns | 0.3166 |     - |     - |    1328 B |

After:

|              Method |                 text |                 hash |                value |             Mean |           Error |          StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------------- |--------------------- |--------------------- |-----------------:|----------------:|----------------:|-------:|------:|------:|----------:|
| PasswordNeedsRehash |                    ? |                    ? |                    ? |         228.8 ns |         1.06 ns |         0.94 ns | 0.0668 |     - |     - |     280 B |