@jvandertil 2020-01-29
With the parser in place, I think I can do better still.