@jvandertil 2020-01-29
There we go, no allocations for `PasswordNeedsRehash`

|              Method |                 text |                 hash |                value |             Mean |         Error |        StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------------- |--------------------- |--------------------- |-----------------:|--------------:|--------------:|-------:|------:|------:|----------:|
| PasswordNeedsRehash |                    ? |                    ? |                    ? |         177.1 ns |       0.09 ns |       0.08 ns |      - |     - |     - |         - |