@jvandertil 2020-01-29
I noticed I was missing support for hashes that follow ‘$2$...’ form. Not sure if important, but currently working that in.