@jvandertil 2020-01-29
Adding that made the parser even slightly faster. I'm guessing the branch predictor does a better job with the more sequential layout of the code.

|              Method |                 text |                 hash |                value |             Mean |        Error |       StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |--------------------- |--------------------- |--------------------- |-----------------:|-------------:|-------------:|-------:|------:|------:|----------:|
| PasswordNeedsRehash |                    ? |                    ? |                    ? |         158.7 ns |      0.10 ns |      0.09 ns |      - |     - |     - |         - |